+++ src/lib/Transport/DataSubscriber.h
@@ -218,9 +218,9 @@ namespace Transport
         void RunDataChannelResponseThread();
 
         // Command channel callbacks
-        void ReadPayloadHeader(const ErrorCode& error, uint32_t bytesTransferred);
-        void ReadPacket(const ErrorCode& error, uint32_t bytesTransferred);
-        void WriteHandler(const ErrorCode& error, uint32_t bytesTransferred);
+        void ReadPayloadHeader(const ErrorCode& error, size_t bytesTransferred);
+        void ReadPacket(const ErrorCode& error, size_t bytesTransferred);
+        void WriteHandler(const ErrorCode& error, size_t bytesTransferred);
 
         // Server response handlers
         void ProcessServerResponse(uint8_t* buffer, uint32_t offset, uint32_t length);
